refactor(SidePage): extract `withContext` helper function to utils
