+++ packages/retail-ui/components/SidePage/SidePageBody.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import styles from './SidePage.less';
-import { SidePageContextType, withContext } from './SidePageContext';
+import { SidePageContextType, SidePageContext } from './SidePageContext';
+import { withContext } from '../../lib/utils';
 
 export interface SidePageBodyProps {
   children?: React.ReactNode;
@@ -20,6 +21,6 @@ export class SidePageBody extends React.Component<SidePageBodyProps> {
   }
 }
 
-export const SidePageBodyWithContext = withContext<SidePageBodyProps>(SidePageBody);
+export const SidePageBodyWithContext = withContext(SidePageContext.Consumer)(SidePageBody);
 
 export default SidePageBodyWithContext;
